Handle multiple powermate devices
This makes our udev config label our powermate symlink with the number
from the input dev it points to, and then on startup builds an event
queue for each of them.  Additionally it changes the script to run so
that the same script ("powermate" in $PATH) is called for any event, and
the invocations are like:

powermate powermate7 left
powermate powermate7 right
powermate powermate7 button 1
powermate powermate7 button 0

(Note the slight change in semantics; button release events were
previously ignored, and are now "0")

This makes no attempt to decide what to /do/ about which device is
which.

Signed-off-by: Peter Jones <pjones@redhat.com>